build: move source files to src/, publish npm package from dist/
