+++ .github/workflows/publish-npm-package.yaml
@@ -17,13 +17,14 @@ jobs:
       - run: npm run lint
       - run: npm run test
       - run: npm run build
-      - run: |
+      - working-directory: ./dist
+        run: |
           npm pack
           mv *.tgz package.tgz
       - uses: actions/upload-artifact@v1
         with:
           name: artifact
-          path: package.tgz
+          path: dist/package.tgz
 
   Publish to npm:
     needs: build
