+++ Dockerfile
@@ -7,10 +7,12 @@ RUN npm run build
 FROM node:alpine
 
 LABEL org.opencontainers.image.source=https://github.com/verkkokauppacom/teams-logger
-ENTRYPOINT [ "/bin/teams-logger.js" ]
+ENTRYPOINT [ "/teams-logger/bin.js" ]
 
-COPY --from=builder /index.js /package.json /package-lock.json .
+RUN mkdir /teams-logger
+WORKDIR /teams-logger
+
+COPY --from=builder /package.json /package-lock.json .
 RUN npm ci --production && rm package-lock.json
 
-COPY --from=builder /bin/*.js ./bin
-COPY --from=builder /lib/*.js ./lib
+COPY --from=builder /dist/* .
