--- index.ts
@@ -1,2 +1,0 @@
-export { default as rawLogger } from './lib/rawLogger'
-export { default as simpleLogger } from './lib/simpleLogger'
