+++ src/simpleLogger.ts
