+++ src/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "outDir": "../dist",
+        "noEmit": false
+    },
+    "exclude": [
+        "../test"
+    ]
+}
